geoprop-ai/backend/requirements.txt
fastapi==0.104.1 
uvicorn==0.24.0 
torch==2.1.0 
scikit-learn==1.3.2 
web3==6.11.3 
httpx==0.25.2 
pyjwt==2.8.0 
cryptography==41.0.7 
geopy==2.4.0 
numpy==1.24.3 
pandas==2.0.3 
joblib==1.3.2 
python-multipart==0.0.6 
python-jose[cryptography]==3.3.0 
